# openapi.yaml
openapi: 3.0.0
info:
  title: Service Envelope API v2.0
  version: 2.1.0
  description: REST API for the Service Envelope System. Includes OTP bearer
    authentication, Mongo-backed auth users and roles, generic email template
    fallbacks, schema v1.0.5 service validation, and unified delivery selection.
  contact:
    name: Service Envelope Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development
  - url: https://api.mapua.edu.ph
    description: Production
paths:
...
  /api/approvals/admin/tokens/{requestId}:
    get:
      tags:
        - Approvals
      summary: Get Approval Tokens (Approvals Admin)
      operationId: getApprovalTokensByRequestInApprovalsRoute
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Tokens for request
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  count:
                    type: integer
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApprovalTokenInfo"
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    requestId: sample
                    count: 1
                    tokens:
                      - token: sample
                        approverId: sample
                        requestId: sample
                        used: true
                        expiresAt: 2026-06-07T00:00:00.000Z
                        createdAt: 2026-06-07T00:00:00.000Z
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET
            "http://localhost:8000/api/approvals/admin/tokens/REQ-20260606-865"
  /api/approvals/{token}:
    get:
      tags:
        - Approvals
      summary: Check Approval Token Status
      operationId: checkApprovalTokenStatus
      security: []
      parameters:
        - $ref: "#/components/parameters/TokenParam"
      responses:
        "200":
          description: Token status
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  requestId:
                    type: string
                  approverId:
                    type: string
                  used:
                    type: boolean
                  expired:
                    type: boolean
                  expiresAt:
                    type: string
                    format: date-time
                  createdAt:
                    type: string
                    format: date-time
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    token: sample
                    requestId: sample
                    approverId: sample
                    used: true
                    expired: true
                    expiresAt: 2026-06-07T00:00:00.000Z
                    createdAt: 2026-06-07T00:00:00.000Z
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET
            "http://localhost:8000/api/approvals/94c2fea5-7de6-427c-9da8-ad3948673177"
  /api/approvals/{token}/approve:
    post:
      tags:
        - Approvals
      summary: Approve Request by Token
      operationId: approveRequest
      security: []
      parameters:
        - $ref: "#/components/parameters/TokenParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  example: Approved by registrar.
            examples:
              sample:
                summary: Sample request
                value:
                  comment: Approved by registrar.
      responses:
        "200":
          description: Approval processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  status:
                    type: string
                    example: approved
                  message:
                    type: string
                  approvalComplete:
                    type: boolean
                  rule:
                    type: string
                  nextStatus:
                    type: string
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    requestId: sample
                    status: approved
                    message: sample
                    approvalComplete: true
                    rule: sample
                    nextStatus: sample
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Approval stage already finalized
      x-codeSamples:
        - lang: bash
          label: cURL
          source: |-
            curl -X POST "http://localhost:8000/api/approvals/94c2fea5-7de6-427c-9da8-ad3948673177/approve" \
              -H "Content-Type: application/json" \
              -d '{"comment":"Approved by registrar."}'
  /api/approvals/{token}/deny:
    post:
      tags:
        - Approvals
      summary: Deny Request by Token
      operationId: denyRequest
      security: []
      parameters:
        - $ref: "#/components/parameters/TokenParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  example: Missing required supporting document.
            examples:
              sample:
                summary: Sample request
                value:
                  reason: Missing required supporting document.
      responses:
        "200":
          description: Request cancelled after denial
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  status:
                    type: string
                    example: cancelled
                  message:
                    type: string
                  reason:
                    type: string
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    requestId: sample
                    status: cancelled
                    message: sample
                    reason: sample
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Approval stage already finalized
      x-codeSamples:
        - lang: bash
          label: cURL
          source: |-
            curl -X POST "http://localhost:8000/api/approvals/94c2fea5-7de6-427c-9da8-ad3948673177/deny" \
              -H "Content-Type: application/json" \
              -d '{"reason":"Missing required supporting document."}'
  /api/approvals/{token}/request:
    get:
      tags:
        - Approvals
      summary: Get Request Details for Approval Token
      operationId: getApprovalTokenRequest
      security: []
      parameters:
        - $ref: "#/components/parameters/TokenParam"
      responses:
        "200":
          description: Request details for approval UI
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    sampleKey: sampleValue
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET
            "http://localhost:8000/api/approvals/94c2fea5-7de6-427c-9da8-ad3948673177/request"
...
  /api/admin/email-templates/{templateId}:
    get:
      tags:
        - Admin
      summary: Get Email Template
      operationId: getEmailTemplate
      security: []
      parameters:
        - $ref: "#/components/parameters/TemplateIdParam"
      responses:
        "200":
          description: Template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    id: SERV-1-approval
                    name: Approval Request Template
                    subject: Approval needed for {{requestId}}
                    htmlBody: sample
                    description: sample
                    templateScope: generic
                    eventKey: sample
                    serviceType: sample
                    isActive: true
                    envelopeType: sample
                    phase: sample
                    createdAt: 2026-06-07T00:00:00.000Z
                    updatedAt: 2026-06-07T00:00:00.000Z
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET
            "http://localhost:8000/api/admin/email-templates/delivery-email-document"
    put:
      tags:
        - Admin
      summary: Update Email Template
      operationId: updateEmailTemplate
      security: []
      parameters:
        - $ref: "#/components/parameters/TemplateIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateUpdate"
            examples:
              sample:
                summary: Sample request
                value:
                  name: sample
                  subject: sample
                  htmlBody: sample
      responses:
        "200":
          description: Template updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: |-
            curl -X PUT "http://localhost:8000/api/admin/email-templates/delivery-email-document" \
              -H "Content-Type: application/json" \
              -d '{"name":"sample","subject":"sample","htmlBody":"sample"}'
    delete:
      tags:
        - Admin
      summary: Delete Email Template
      operationId: deleteEmailTemplate
      security: []
      parameters:
        - $ref: "#/components/parameters/TemplateIdParam"
      responses:
        "200":
          description: Template deleted
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X DELETE
            "http://localhost:8000/api/admin/email-templates/delivery-email-document"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RequestIdParam:
      name: requestId
      in: path
      required: true
      schema:
        type: string
      example: REQ-20260606-042
      description: Request identifier.
    TokenParam:
      name: token
      in: path
      required: true
      schema:
        type: string
      example: 2d2f5e0a-5a8f-4b7d-a365-6dd77c9f17d8
      description: Approval or feedback token.
...
    TemplateIdParam:
      name: templateId
      in: path
      required: true
      schema:
        type: string
      example: SERV-1-approval
      description: Email template ID.
...
  schemas:
...
    ApprovalTokenInfo:
      type: object
      properties:
        token:
          type: string
        approverId:
          type: string
        requestId:
          type: string
        used:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    EmailTemplate:
      type: object
      properties:
        id:
          type: string
          example: SERV-1-approval
        name:
          type: string
          example: Approval Request Template
        subject:
          type: string
          example: Approval needed for {{requestId}}
        htmlBody:
          type: string
        description:
          type: string
        templateScope:
          type: string
          enum:
            - generic
            - envelope
            - service
        eventKey:
          type: string
        serviceType:
          type: string
        isActive:
          type: boolean
        envelopeType:
          type: string
        phase:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    EmailTemplateUpdate:
      type: object
      required:
        - name
        - subject
        - htmlBody
      properties:
        name:
          type: string
        subject:
          type: string
        htmlBody:
          type: string
        description:
          type: string
        templateScope:
          type: string
          enum:
            - generic
            - envelope
            - service
        eventKey:
          type: string
        serviceType:
          type: string
        isActive:
          type: boolean
        envelopeType:
          type: string
        phase:
          type: string
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Request not found
        details:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 400 response
              value:
                error: Request not found
                details: sample
...
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 404 response
              value:
                error: Request not found
                details: sample
...